add location store to store retrieved location data and also added logic to inject location data in headers

+++ src/main/kotlin/model/LocationReq.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class LocationReq(
     @SerialName("horizontal_accuracy")
-    val horizontalAccuracy: Int,
+    val horizontalAccuracy: Double,
     @SerialName("is_precise_location_denied")
     val isPreciseLocationDenied: Boolean,
     @SerialName("lon")
